requirements.txt
beautifulsoup4>=4.12.0
networkx>=3.1
requests>=2.31.0
pytest>=7.0.0
pyyaml>=6.0.0
python-dotenv
mcp